Update to UI, layout, and animations
Added a standard menu bar, swapping out to a hamburger menu on smaller screens. Added slide and fade-in animations to banner and main sections. Adjusted layout to use full height of page and place footer at the bottom to keep the footer at the bottom of the screen even when little content is shown in banner or main.